sql bigint -> int

+++ db/isubata.sql
@@ -1,5 +1,5 @@
 CREATE TABLE user (
-  id BIGINT UNSIGNED AUTO_INCREMENT NOT NULL PRIMARY KEY,
+  id INT UNSIGNED AUTO_INCREMENT NOT NULL PRIMARY KEY,
   name VARCHAR(191) UNIQUE,
   salt VARCHAR(20),
   password VARCHAR(40),
@@ -10,13 +10,13 @@ CREATE TABLE user (
 ) Engine=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE image (
-  id BIGINT UNSIGNED AUTO_INCREMENT NOT NULL PRIMARY KEY,
+  id INT UNSIGNED AUTO_INCREMENT NOT NULL PRIMARY KEY,
   name VARCHAR(191),
   data LONGBLOB
 ) Engine=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE channel (
-  id BIGINT AUTO_INCREMENT NOT NULL PRIMARY KEY,
+  id INT AUTO_INCREMENT NOT NULL PRIMARY KEY,
   name TEXT NOT NULL,
   description MEDIUMTEXT,
   updated_at DATETIME NOT NULL,
@@ -24,9 +24,9 @@ CREATE TABLE channel (
 ) Engine=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE message (
-  id BIGINT AUTO_INCREMENT NOT NULL PRIMARY KEY,
-  channel_id BIGINT,
-  user_id BIGINT,
+  id INT AUTO_INCREMENT NOT NULL PRIMARY KEY,
+  channel_id INT,
+  user_id INT,
   content TEXT,
   created_at DATETIME NOT NULL,
   count INT,
@@ -34,9 +34,9 @@ CREATE TABLE message (
 ) Engine=InnoDB DEFAULT CHARSET=utf8mb4;
 
 CREATE TABLE haveread (
-  user_id BIGINT NOT NULL,
-  channel_id BIGINT NOT NULL,
-  message_id BIGINT,
+  user_id INT NOT NULL,
+  channel_id INT NOT NULL,
+  message_id INT,
   updated_at DATETIME NOT NULL,
   created_at DATETIME NOT NULL,
   PRIMARY KEY(user_id, channel_id)
